requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
cryptography==41.0.7
pytest==7.4.3
python-dotenv==1.0.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
pytz==2023.3


